Move Platform class into utils.
