+++ mujoco_toolbox/Utils.py
@@ -1,7 +1,10 @@
 from functools import wraps
-from sys import platform
 from time import time
 from colorama import Fore, Style, init
+import platform
+import sys
+from multiprocessing import cpu_count
+from screeninfo import get_monitors
 
 
 def timer(func):
@@ -20,7 +23,6 @@ def timer(func):
 
     return time_wrapper
 
-
 def print_warning(*args, prefix: bool = True) -> None:
     """
     Prints the first argument as an orange warning message and prints the other arguments in regular text, all on the same line.
@@ -36,7 +38,7 @@ def print_warning(*args, prefix: bool = True) -> None:
             message = args[0]
 
         # Check if we're on Windows or a Unix-like system for color support
-        if platform.startswith("win"):
+        if sys.platform.startswith("win"):
             try:
                 init()  # Initialize colorama
                 print(f"{Fore.YELLOW}{message}{Style.RESET_ALL}", end=" ")  # Yellow color for warning
@@ -65,7 +67,7 @@ def print_success(*args, prefix: bool = True) -> None:
             message = args[0]
 
         # Check if we're on Windows or a Unix-like system for color support
-        if platform.startswith("win"):
+        if sys.platform.startswith("win"):
             try:
                 init()  # Initialize colorama
                 print(f"{Fore.GREEN}{message}{Style.RESET_ALL}", end=" ")  # Green color for success
@@ -78,3 +80,49 @@ def print_success(*args, prefix: bool = True) -> None:
             # ANSI escape code for green (for Unix-like systems)
             print(f"\033[32m{message}\033[0m", end=" ")  # Green ANSI escape code
             print(*args[1:])
+
+class _Platform:
+    def __init__(self):
+        self.OS = platform.system()
+        self.PROCESSOR = platform.processor()
+        self.MACHINE = platform.machine()
+        self.PYTHON_VERSION = platform.python_version()
+        self.IDE = self.get_developer_mode()
+        self.CPU_COUNT = cpu_count()
+        self.ARCHITECTURE = platform.architecture()[0]
+        self.PYTHON_IMPLEMENTATION = platform.python_implementation()
+        self.RESOLUTION, self.GUI_ENABLED = self.get_resolution()
+        # self.NUM_MONITORS = get_monitors().__len__()
+
+    def __repr__(self):
+        return (
+            f"Operating System: {self.OS}\n"
+            f"Processor: {self.PROCESSOR}\n"
+            f"Machine Type: {self.MACHINE}\n"
+            f"Python Version: {self.PYTHON_VERSION}\n"
+            f"IDE: {self.IDE}\n"
+            f"CPU Count: {self.CPU_COUNT}\n"
+            f"Architecture: {self.ARCHITECTURE}\n"
+            f"Python Implementation: {self.PYTHON_IMPLEMENTATION}\n"
+            f"Resolution (disp0): {self.RESOLUTION}\n"
+            f"GUI Enabled: {self.GUI_ENABLED}"
+        )
+    
+    def __str__(self):
+        return self.__repr__()
+
+    @staticmethod
+    def get_developer_mode():
+        if "ipykernel" in sys.modules and not sys.stdin.isatty():
+            return "jupyter"
+        return "terminal" if hasattr(sys, "ps1") else "script"
+    
+    @staticmethod
+    def get_resolution():
+        """Detects screen resolution and potential headless operation."""
+        try:
+            monitor = get_monitors()[0]
+            return (monitor.width, monitor.height), True
+        except Exception:
+            print_warning("Detected headless operation. Disabling GUI...")
+            return (1920, 1080), False
